app_uifce_back/djangoRest_back/serializers/EstadoSerializer.py
from rest_framework import serializers
from djangoRest_back.models import Estado

class EstadoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Estado
        fields = ['nombre']